Add heze test exercising non-default inclusion_list_bits in bid
Extracts `prepare_signed_execution_payload_bid` and
`run_execution_payload_bid_processing` from
`gloas/block_processing/test_process_execution_payload_bid.py` into a new
shared module `helpers/execution_payload_bid.py`. The preparation helper
gains an `inclusion_list_bits` kwarg that is only applied post-heze.

The heze test
`test_process_execution_payload_bid_valid_builder_with_non_default_inclusion_list_bits`
sets one bit before signing, forcing the heze SSZ merkle root to diverge
from the gloas root. Without this, an implementation that incorrectly
hashes a heze bid using the gloas `ExecutionPayloadBid` container still
produces the right root (the extra leaf collides with gloas's zero
padding), so existing cascaded tests cannot catch that class of bug.
